image/bot/requirements.txt
python-dotenv
discord.py
aiosqlite
aiohttp
cogwatch